[jaypedia/airbnb#19] Rename: 폴더명 복수형으로 변경
- util -> utils
- api -> apis
